fix project status update bug
With Pusher integration, Shipscope only polls for a fresh list of
projects every 5 minutes. This bug prevented Shipscope from updating a
project’s status after the project list polling. The cause was that the
project object that the Pusher-update callback was updating was no
longer in the projects collection. It had been replaced by a fresh copy
of the project.

This is fixed by finding the project in the projects collection by
projectId, and updating that project object. This is much more reliable.